server/requirements.txt
pip==9.0.1
Django==1.10.2
django-cors-headers==1.3.1
django-extensions==1.7.4
djangorestframework==3.4.7
psycopg2==2.6.1
whitenoise==3.2.2
coverage==4.2